Extract K3OSConfigReconciler setup code into a separate file
