+++ controllers/config/k3osconfig_controller.go
@@ -22,38 +22,22 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-package controllers
+package config
 
 import (
 	"context"
-	"os"
-	"time"
 
 	configv1alpha1 "github.com/annismckenzie/k3os-config-operator/apis/config/v1alpha1"
-	"github.com/go-logr/logr"
 	"k8s.io/apimachinery/pkg/api/errors"
-	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
 	ctrl "sigs.k8s.io/controller-runtime"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/manager"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
-	"sigs.k8s.io/controller-runtime/pkg/runtime/inject"
 )
 
 // response is a helper struct to cut down on the amount of if and switch statements.
 type response struct {
 	result reconcile.Result
 	err    error
-}
-
-// K3OSConfigReconciler reconciles a K3OSConfig object.
-type K3OSConfigReconciler struct {
-	client                 client.Client
-	logger                 logr.Logger
-	scheme                 *runtime.Scheme
-	leader                 bool
-	defaultRequeueResponse ctrl.Result
 }
 
 // +kubebuilder:rbac:groups=config.operators.annismckenzie.github.com,resources=k3osconfigs,verbs=get;list;watch;create;update;patch;delete
@@ -99,74 +83,4 @@ func (r *K3OSConfigReconciler) fetchK3OSConfig(ctx context.Context, name types.N
 		return nil, &response{result: r.defaultRequeueResponse, err: nil}, err
 	}
 	return config.DeepCopy(), nil, nil
-}
-
-// Option denotes an option for configuring this controller.
-type Option interface{}
-
-type requireLeaderElectionOpt struct {
-	requireLeaderElection bool
-}
-
-// RequireLeaderElection returns an option that requires the operator being the leader to run this controller instance.
-func RequireLeaderElection() Option {
-	return &requireLeaderElectionOpt{}
-}
-
-// https://github.com/kubernetes-sigs/controller-runtime/pull/921#issuecomment-662187521 doesn't work
-// but there's always another way. 🥁 🥁 🥁
-type nonLeaderLeaseNeedingManagerWrapper struct {
-	manager.Manager
-}
-
-func (w *nonLeaderLeaseNeedingManagerWrapper) Add(r manager.Runnable) error {
-	return w.Manager.Add(&nonLeaderLeaseNeedingRunnableWrapper{Runnable: r})
-}
-
-type nonLeaderLeaseNeedingRunnableWrapper struct {
-	manager.Runnable
-}
-
-// NeedLeaderElection satisfies manager.LeaderElectionRunnable interface.
-func (w *nonLeaderLeaseNeedingRunnableWrapper) NeedLeaderElection() bool {
-	return false
-}
-
-// SetupWithManager is called in main to setup the K3OSConfig reconiler with the manager as a non-leader.
-func (r *K3OSConfigReconciler) SetupWithManager(mgr ctrl.Manager, options ...Option) error {
-	r.defaultRequeueResponse = ctrl.Result{RequeueAfter: time.Second * 30}
-
-	for _, option := range options {
-		if _, ok := option.(*requireLeaderElectionOpt); ok {
-			r.leader = true
-		}
-	}
-
-	if !r.leader { // wrap manager so this can run without a leader lease
-		mgr = &nonLeaderLeaseNeedingManagerWrapper{Manager: mgr}
-	}
-
-	// cannot inject via inject.LoggerInto because `leader` field isn't set at that point
-	r.logger = mgr.GetLogger().
-		WithName("controllers").
-		WithName("K3OSConfig").
-		WithValues("podName", os.Getenv("HOSTNAME"), "leader", r.leader)
-
-	return ctrl.NewControllerManagedBy(mgr).For(&configv1alpha1.K3OSConfig{}).Complete(r)
-}
-
-// Interface implementations for dependency injection
-var _ inject.Client = (*K3OSConfigReconciler)(nil)
-var _ inject.Scheme = (*K3OSConfigReconciler)(nil)
-
-// InjectClient satisfies the inject.Client interface.
-func (r *K3OSConfigReconciler) InjectClient(client client.Client) error {
-	r.client = client
-	return nil
-}
-
-// InjectScheme satisfies the inject.Scheme interface.
-func (r *K3OSConfigReconciler) InjectScheme(scheme *runtime.Scheme) error {
-	r.scheme = scheme
-	return nil
 }
